% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_GLM.R
\name{apply_bGLM}
\alias{apply_bGLM}
\title{Run a GLM ...}
\usage{
apply_bGLM(
  sce,
  coef_test = "group_idB",
  formula = ~batch_cov + group_id,
  BPPARAM = BPPARAM,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{coef_test}{Coefficient to test.}

\item{formula}{Design formula.}

\item{BPPARAM}{Parallel execution (ignored for edgeR functions)}

\item{...}{Ignored.}
}
\value{
Results data.frame.
}
\description{
Run a GLM ...
}
