% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_edgeR.R
\name{apply_edgeR_QP}
\alias{apply_edgeR_QP}
\title{Run a GLM ...}
\usage{
apply_edgeR_QP(
  sce,
  coef_test = "group_idB",
  formula = ~batch_cov + group_id,
  BPPARAM,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{coef_test}{Coefficient to test.}

\item{formula}{Design formula.}

\item{BPPARAM}{Parallel execution (ignored for edgeR functions)}

\item{...}{Ignored.}
}
\value{
Results data.frame.
}
\description{
Run a GLM ...
}
