% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_muscat.R
\name{apply_muscat}
\alias{apply_muscat}
\title{Apply muscat with default parameters}
\usage{
apply_muscat(
  sce,
  coef = "group_idB",
  formula = ~batch_cov + group_id,
  subject_id = "subject_id",
  group_id = "group_id",
  cluster_id = "cluster_id",
  ...,
  combined = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{coef}{Coefficient to test.}

\item{formula}{Design formula.}

\item{subject_id}{Name of the \code{colData} column containg subject identifiers.}

\item{group_id}{Name of the \code{colData} column containg group identifiers.}

\item{cluster_id}{Name of the \code{colData} column containg cluster identifiers.}

\item{...}{Ignored.}

\item{combined}{Logical. Should the pseudobulk approach be applied on
multiple celltypes together?. Default: \code{FALSE}.}
}
\value{
Results data.frame.
}
\description{
When analysing multiple celltypes combined, uses a workaround with
\code{scran::pseudobulkDGE()} to apply the pseudobulk DE approach.
\code{edgeR::glmQLFTest()} is used as the underlying DE method.
}
