% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ttest.R
\name{apply_ttest}
\alias{apply_ttest}
\title{Apply regular pairwise t-tests}
\usage{
apply_ttest(sce, ..., group_id = "group_id", batch_id = "batch_cov")
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{...}{Ignored.}

\item{group_id}{Name of the \code{colData} column containg group identifiers.}

\item{batch_id}{Name of the \code{colData} column containg batch identifiers.}
}
\value{
Results data.frame.
}
\description{
Perform DE analyses using pairwise t-tests as implemented in
\code{\link[scran:pairwiseTTests]{scran::pairwiseTTests()}}, blocking on batch covariates.
}
