% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createMockGroups}
\alias{createMockGroups}
\title{Create subject-level mock groups}
\usage{
createMockGroups(
  x,
  labs = c("A", "B"),
  subject_id = "ind_cov",
  batch_id = "batch_cov"
)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object.}

\item{labs}{Mock labels to use. The length determines the number of groups.}

\item{subject_id}{\code{colData} column name that contains the subject labels.}

\item{batch_id}{\code{colData} column name that contains batch labels.}
}
\value{
An object of class \code{class(x)} with mock labels added in the \code{"mock_group"}
column of the \code{colData}.
}
\description{
Randomly assign subjects to mock groups for a single-cell experiment.
}
