% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes for mock analyses}
\usage{
filterGenes(sce, group, min.count = 1L, min.total.count = 10L, ...)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object.}

\item{group}{vector or factor giving group membership for a oneway layout, if appropriate.}

\item{min.count}{numeric. Minimum count required for at least some samples.}

\item{min.total.count}{numeric. Minimum total count required.}

\item{...}{
  Arguments passed on to \code{\link[edgeR:filterByExpr]{edgeR::filterByExpr}}
  \describe{
    \item{\code{y}}{matrix of counts, or a \code{DGEList} object, or a \code{SummarizedExperiment} object.}
  }}
}
\value{
A \linkS4class{SingleCellExperiment} object with filtered genes removed.
}
\description{
Uses \code{edgeR::filterByExpr()} to remove lowly expressed genes, taking the
design of the experiment into account.
}
