% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMultiSubjectSCE.R
\name{mockMultiSubjectSCE}
\alias{mockMultiSubjectSCE}
\title{Mock up a multi-subject SingleCellExperiment}
\usage{
mockMultiSubjectSCE(
  ncells = 200,
  ngenes = 2000,
  nsubjects = 8,
  ngroups = 2,
  nbatches = 1,
  ...
)
}
\arguments{
\item{ncells}{The desired total number of cells.}

\item{ngenes}{The desired total number of genes.}

\item{nsubjects}{The desired total number of subjects.}

\item{ngroups}{The desired total number of \strong{subject-level} groups.}

\item{nbatches}{The desired total number of \strong{subject-level} batches.}

\item{...}{Further arguments passed on to \code{\link[scuttle:mockSCE]{scuttle::mockSCE()}}.}
}
\value{
A \linkS4class{SingleCellExperiment} object. Subject, group and batch labels
are stored in the \code{"subject_id"}, \code{"mock_group"} and \code{"batch_id"} columns of
the \code{colData}, respectively. Subjects will be equally divided among the
groups.
}
\description{
Calls \code{\link[scuttle:mockSCE]{scuttle::mockSCE()}} and adds subject and group labels.
}
