% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print_plots}
\alias{print_plots}
\title{Create subheading for each celltype and its corresponding plot}
\usage{
print_plots(plotlist, use_names = NULL, header_level = 3)
}
\arguments{
\item{plotlist}{List of plots. Can be nested.}

\item{use_names}{Names to be used in the subheading. If not provided, will
use \code{names(plotlist)}.}

\item{header_level}{Level of the subheading to use. Default: 3.}
}
\description{
For use in RMarkdown code chunks with option \code{results="asis"}.
}
