% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleCellsPerSubject}
\alias{sampleCellsPerSubject}
\title{Sub-sample cells per subject}
\usage{
sampleCellsPerSubject(
  x,
  n_cells,
  subject_id = "ind_cov",
  group_id = "mock_group"
)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object.}

\item{n_cells}{Number of cells per subject to keep. Possible to supply more
than one value.}

\item{subject_id}{\code{colData} column name containing the subject labels}

\item{group_id}{\code{colData} column name containing the group labels. Used for
gene filtering with \code{filterGenes()}.}
}
\value{
A list of objects of class \code{class(x)} with length \code{length(n_cells)}.
Each cobject coresponds to a down-sampled dataset
}
\description{
Down-samples the number of cells for each subject to a given number. Then
calls \code{\link{filterGenes}} on each subset.
}
