% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DE.R
\name{simulate_DE}
\alias{simulate_DE}
\title{Simulate DE by swapping genes}
\usage{
simulate_DE(x, groups, prop_DE = 0.01)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.}

\item{groups}{A vector of length equal to \code{ncol(x)} specifying the group to
which each cell is assigned. DE will be induced between these groups.}

\item{prop_DE}{Numeric scalar specifying the proportion of genes that will be
simulated as DE. Default: 1\%.}
}
\value{
A \linkS4class{SingleCellExperiment} object with DE induced between the
specified \code{groups}. The simulated counts are contained within the \code{"counts"}
assay of the returned object. The original \code{colData(x)} and \code{rowData(x)} are
retained with additional columns in the latter:
\itemize{
\item \code{"is_DE"}: logical vector indicating the ground truth status of each gene
\item \code{"swapped_gene"}: character vector indicating the original gene with which
the gene was swapped
}

Note that normalized values in \code{logcounts(x)} are not retained. For simple
library size-factor normalization, size factors for each cell will be
identical before and after simulation. However, for more sophisticated
methods, such as normalization by deconvolution, the size factors might
differ due to the gene swapping.
}
\description{
Simple DE simulator that takes a dataset with a grouping factor and induces
DE by swapping some genes in one of the groups but not the other. This should
ensure keeping the same data structure as the original data without having to
estimate distribution parameters or making assumptions.
}
\examples{
set.seed(123)
sce <- mockMultiSubjectSCE()

sim_sce <- simulate_DE(sce, groups = sce$mock_group, prop_DE = 0.1)

}
